Refactor category schema methods and added a base model class
