+++ recipe-service/app/models/category.js
@@ -1,5 +1,7 @@
 const mongoose = require('mongoose');
 
+const Model = require('./');
+
 mongoose.Promise = global.Promise;
 const Schema = mongoose.Schema;
 
@@ -31,60 +33,8 @@ const CategorySchema = new Schema({
 /**
  * @Class
  */
-class Category {
-  /**
-   * @param {string} id
-   * @return {object} returns a user
-   */
-  static get(id) {
-    return this.findById(id);
-  }
-
-  /**
-   * Gets all data set, handles pagination and 
-   * query params
-   * @param {int} limit
-   * @param {int} page
-   * @param {q} q user query 
-   * @return {object} returns an object
-   */
-  static getAll(limit = 10, page = 1, q) {
-    const query = q ? { description: new RegExp(q) } : {};
-    const skip = limit * (page - 1);
-
-    const result = {};
-    return new Promise((resolve, reject) => {
-      this.find(query).limit(limit).skip(skip)
-        .then(categories => {
-          result.data = categories;
-          this.count().then(count => {
-            result.count = count;
-            resolve(result);
-          }).catch(error => reject(error));
-        })
-        .catch(error => reject(error));
-    });
-  }
-
-  /**
-   * @param {int} id
-   * @return {object} returns an object
-   */
-  static delete(id) {
-    return this.findByIdAndRemove(id);
-  }
-
-  /**
-   * @param {int} id
-   * @return {object} returns an object
-   */
-  static updateData(id, details) {
-    const options = {
-      runValidators: true,
-      new: true
-    };
-    return this.findByIdAndUpdate(id, details, options);
-  }
+class Category extends Model {
+  // TODO: implement methods related to this schema
 }
 
 CategorySchema.loadClass(Category);
